backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Speech Recognition and Text-to-Speech
torch==2.1.2
torchvision==0.16.2
openai-whisper==20231117
gtts==2.5.1
pydub==0.25.1
speechrecognition==3.10.1
numpy==1.26.3
edge-tts==6.1.9
azure-cognitiveservices-speech==1.35.0
bark==1.0.0
transformers==4.36.0
scipy==1.11.4

# Translation
translate==3.6.1
langdetect==1.0.9

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# AWS
boto3==1.34.21
botocore==1.34.21

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
aiofiles==23.2.1
httpx==0.26.0
